backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
pytest
httpx
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart